Also changed floats to doubles in the gtests.

+++ cpp/gtest/TestMatrixResizes.cpp
@@ -1,13 +1,13 @@
 #include <gtest/gtest.h>
 #include <Eigen/Dense>
 
-using Eigen::MatrixXf;
-using Eigen::Matrix2f;
-using Eigen::Matrix3f;
+using Eigen::MatrixXd;
+using Eigen::Matrix2d;
+using Eigen::Matrix3d;
 
 TEST(FASTSLAM_TEST, basic_matrix_assign) {
-  MatrixXf m(2,2);
-  m = MatrixXf::Ones(2,2);
-  EXPECT_NO_THROW(Matrix2f m2 = m);
-  EXPECT_DEATH(Matrix3f m3 = m, "Invalid sizes when resizing a matrix or array.");
+  MatrixXd m(2,2);
+  m = MatrixXd::Ones(2,2);
+  EXPECT_NO_THROW(Matrix2d m2 = m);
+  EXPECT_DEATH(Matrix3d m3 = m, "Invalid sizes when resizing a matrix or array.");
 }
